+++ src/App.tsx
@@ -1,30 +1,50 @@
 import { Routes, Route } from "react-router-dom";
 import Main from "../src/app/Main";
 import AccountContent from "../src/app/AccountContent";
+import { TitleTypes } from "../src/app/types";
 
 import "./App.css";
 
 function App(): JSX.Element {
-  const accountIndex = {
-    accountOne: 1,
-    accountTwo: 2,
-    accountThree: 3,
-  };
+
+  const accounts: TitleTypes[] = [
+    {
+      key: "1",
+      id: "1",
+      name: "Account 1",
+      createdOn: "5-8-2021",
+      owner: "User 1",
+    },
+    {
+      key: "2",
+      id: "2",
+      name: "Account 2",
+      createdOn: "5-8-2021",
+      owner: "User 2",
+    },
+    {
+      key: "3",
+      id: "3",
+      name: "Account 3",
+      createdOn: "5-8-2021",
+      owner: "User 3",
+    },
+  ];
   return (
     <div className="App">
       <Routes>
-        <Route path="/" element={<Main />} />
+        <Route path="/" element={<Main accounts={accounts}/>} />
         <Route
           path="/account1"
-          element={<AccountContent accountNumber={accountIndex.accountOne} />}
+          element={<AccountContent accountNumber={accounts[0].id} />}
         />
         <Route
           path="/account2"
-          element={<AccountContent accountNumber={accountIndex.accountTwo} />}
+          element={<AccountContent accountNumber={accounts[1].id} />}
         />
         <Route
           path="/account3"
-          element={<AccountContent accountNumber={accountIndex.accountThree} />}
+          element={<AccountContent accountNumber={accounts[2].id} />}
         />
       </Routes>
     </div>
