+++ src/app/AccountContent.tsx
@@ -5,7 +5,7 @@ export default function AccountContent({ accountNumber }: any): JSX.Element {
     <div className="account">
       <h1>
         Account{" "}
-        {accountNumber === 2 ? "Two" : accountNumber === 1 ? "One" : "Three"}
+        {accountNumber === "2" ? "Two" : accountNumber === "1" ? "One" : "Three"}
       </h1>
       <NavLink to="/">Back to home</NavLink>
     </div>
