+++ src/app/style.scss
@@ -40,6 +40,7 @@
   justify-content: space-around;
   width: 100%;
   padding: 50px;
+
   a {
     text-decoration: none;
     border: 1px solid rgb(56, 125, 255);
